qt4-native: introduce the native version of qt4 and remove old qt4-tools-native
The new recipe is based on the qt4-tool-native one but it even installs the libraries of
qt4 and not only the tools like moc/rcc. It will not increase the build time as most
libraries are already build for the qt4-tool-native but not staged.

This commit removes the old qt4-tools-native too. All versions provided by
qt4-tools-native are even supported by qt4-native so it is a drop-in replacement.

Signed-off-by: Simon Busch <morphis@gravedo.de>
Acked-by: Otavio Salvador <otavio@ossystems.com.br>